Prevent invalid filter being set for Pcaplet
I think this was working for Metalware previously as after parsing the
options string given, `ARGV` would be an empty string, and so the filter
would be set to an empty string which would later be overridden.

However, for the new Metalware CLI we now run the CLI directly rather
than re-executing another script, and so `ARGV` will not be empty but
rather will contain the command name (`hunter`). This is not a valid
filter but `Pcaplet` would attempt to use it as one, which would cause
an error.

Rather than modifying `ARGV` and then attempting to use that to get the
filter (which `Pcaplet` does apparently to support both use as a library
and a script), let's now just set filters via `add_filter`; which is
clearer, prevents the error, and is what we actually do in practise in
the existing code.